refactor(desktop): split agents page orchestration flow
